Fix pricing page metadata for client components
